refactor: 예외 처리 기능을 포함한 입력 로직 유틸 클래스 인터페이스 제거
