+++ src/main/kotlin/LottoGameApplication.kt
@@ -1,6 +1,6 @@
 import controller.LottoGameController
 import model.RandomNumbersGenerator
-import util.ExceptionHandlerImpl
+import util.ExceptionHandler
 import view.ConsoleLottoGameInputView
 import view.ConsoleLottoGameOutputView
 
@@ -8,7 +8,7 @@ fun main() {
     LottoGameController(
         ConsoleLottoGameInputView(),
         ConsoleLottoGameOutputView(),
-        ExceptionHandlerImpl(),
+        ExceptionHandler(),
         RandomNumbersGenerator(),
     ).startLottoGame()
 }
